remove unneeded CRC calculations (#50)
* remove unneeded CRC calculations

file_name previously used file_info as a filter for "good" files, i.e., to
exclude non-files from a directory listing.   It also calculated
the file crc.  This seemed useful
enough initially: corrupt files would
be excluded.   However, it it a big
drag on large (100M+) files, because
the CRC ic calculated every time, which is slow and not necessary,
because if the files are corrupt they will be detected elsewhere.

* log reqeuest parsing error and include in 4xx response

* update detected version

* depend on VERSION